+++ TrackOrJargh/src/main/java/com/trackorjargh/javaclass/PointShow.java
@@ -20,12 +20,12 @@ public class PointShow {
 	@OneToOne
 	private User user;
 	
-	private Long points;
+	private double points;
 
 	public PointShow() {
 	}
 
-	public PointShow(Long points) {
+	public PointShow(double points) {
 		this.points = points;
 	}
 
@@ -53,11 +53,11 @@ public class PointShow {
 		this.user = user;
 	}
 
-	public Long getPoints() {
+	public double getPoints() {
 		return points;
 	}
 
-	public void setPoints(Long points) {
+	public void setPoints(double points) {
 		this.points = points;
 	}
 }
